requirements.txt
readability-lxml
requests
bottle
